refactor: moved getJobParameters in JobParametersDB
